+++ src/rexecop/plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Trusted in-process plugin host contracts."""